requirements.txt
aiogram>=3.4.1
feedparser>=6.0.11
python-dotenv>=1.0.1
apscheduler>=3.10.4
aiofiles>=23.2.1
google-genai>=0.2.0
